requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
jupyter
gymnasium
stable-baselines3
torch
streamlit
google-generativeai